HUD 구성 완료(Jul 22, 2022) TODO: UI 배치 Object 배치 File정리

+++ WinApi2D/CScene_Ending.cpp
@@ -90,7 +90,7 @@ void CScene_Ending::Enter()
 	int cel;
 	temp = CScene_Start::timer;
 	min = temp / 60;
-	sec = ceil((short)temp % 60);
+	sec = ceil((int)temp % 60);
 	cel = ceil((int)(temp * 100) % 100);
 	wstring Time_min;
 	wstring Time_sec;
